backend/requirements.txt
asgiref
atlastk
cffi
click
cryptography
Cython
ecdsa
fastapi
FastAPI-SQLAlchemy
greenlet
h11
joblib
numpy
pyasn1
pycparser
pydantic
python-jose
python-multipart
rsa
scikit-learn
scipy
six
SQLAlchemy
starlette
threadpoolctl
tortoise
typing-extensions
uvicorn
passlib
bcrypt